% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBreakpoints}
\alias{plotBreakpoints}
\title{Plotting function for BreakPointR}
\usage{
plotBreakpoints(data, plotLibraries = NULL, file = NULL)
}
\arguments{
\item{output}{genomic ranges from the function runBreakpointr or runBreakpointrALL}
}
\description{
Plotting function for BreakPointR
}
\author{
David Porubsky
}

