% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBedFile.R
\name{writeBedFile}
\alias{writeBedFile}
\title{Generates a bedfile from inFile}
\usage{
writeBedFile(fileName, dataDirectory, fragments = NULL, deltaWs = NULL,
  breakTrack = NULL, bin = NA, col = "103,139,139 243,165,97")
}
\arguments{
\item{fileName}{is used to index the bedfile(s)}

\item{dataDirectory}{is location to write bedfile(s)}

\item{fragments}{A \code{\link{GRanges}} object with strand and mapq metadata, such as that generated by \code{\link{bam2GRanges}}}

\item{deltaWs}{A \code{\link{GRanges}} object with metadata column "deltaW" generated by \code{\link{deltaWCalculator}}}

\item{breakTrack}{A \code{\link{GRanges}} object with metadata "genoT" (e.g. newBreaks) will write a bedtrack with refined breakpoints}
}
\description{
Write a bedfile from Breakpoint.R files for upload on to UCSC Genome browser
}
\author{
Ashley Sanders, David Porubsky, Aaron Taudt
}

